convert tailwind theme into typescript
